requirements.txt
google-cloud-storage
google-cloud-secret-manager